use docker version 1.10

+++ scripts/end2end_test.sh
@@ -2,15 +2,23 @@
 
 set -x # debugging
 
-DOCKER_COMPOSE_EXEC="/tmp/docker-compose"
-DOCKER_COMPOSE_FILE="docker/docker-compose.v1.yml"
+# settings
+DOCKER_COMPOSE_CONFIG_V1=${DOCKER_COMPOSE_CONFIG_V1:-0}
 DOCKER_COMPOSE_VERSION="1.8.1"
+
+DOCKER_COMPOSE_EXEC="/tmp/docker-compose"
+DOCKER_COMPOSE_FILE="docker/docker-compose.v2.yml"
 curl -L "https://github.com/docker/compose/releases/download/${DOCKER_COMPOSE_VERSION}/docker-compose-Linux-x86_64" > ${DOCKER_COMPOSE_EXEC}
 chmod +x ${DOCKER_COMPOSE_EXEC}
 
+if [ ${DOCKER_COMPOSE_CONFIG_V1} -eq 1 ]
+then
+  DOCKER_COMPOSE_FILE="docker/docker-compose.v1.yml"
+  export DOCKER_NETWORK="docker-net"
+  docker network create --driver bridge ${DOCKER_NETWORK}
+fi
+
 # start all docker containers
-export DOCKER_NETWORK="docker-net"
-docker network create --driver bridge ${DOCKER_NETWORK}
 ${DOCKER_COMPOSE_EXEC} -f ${DOCKER_COMPOSE_FILE} up -d
 
 # wait for carbon input before sending data
@@ -27,7 +35,12 @@ scripts/verify_metrics_received.py 127.0.0.1 8080 10 ${METRICS_PER_SECOND}
 RESULT=${?}
 
 ${DOCKER_COMPOSE_EXEC} -f ${DOCKER_COMPOSE_FILE} down
-docker network rm ${DOCKER_NETWORK}
+
+if [ ${DOCKER_COMPOSE_CONFIG_V1} -eq 1 ]
+then
+  docker network rm ${DOCKER_NETWORK}
+fi
+
 scripts/generate_test_data.sh stop
 
 exit  ${RESULT}
